fix: UI update
